+++ grammar.js
@@ -251,30 +251,17 @@ module.exports = grammar({
     )),
     _module_path: $ => repeat1($.module_resolution),
     path_ident: $ => seq(optional($._module_path), $.ident),
+    path_at_ident: $ => seq(optional($._module_path), $.at_ident),
     path_type_ident: $ => seq(optional($._module_path), $.type_ident),
     path_const_ident: $ => seq(optional($._module_path), $.const_ident),
     path_at_type_ident: $ => seq(optional($._module_path), $.at_type_ident),
-    alias_path_ident: $ => seq(
-      optional($._module_path),
-      optional(
-        seq(
-          $.type_ident,
-          optional($.generic_arguments),
-          '.'
-        ),
-      ),
-      $.ident,
+    alias_path_ident: $ => choice(
+      seq($.type, '.', $.ident),
+      $.path_ident,
     ),
     alias_path_at_ident: $ => seq(
-      optional($._module_path),
-      optional(
-        seq(
-          $.type_ident,
-          optional($.generic_arguments),
-          '.'
-        ),
-      ),
-      $.at_ident,
+      seq($.type, '.', $.at_ident),
+      $.path_at_ident,
     ),
 
     // Generic Parameters